relocate km_selector
